Address PR comments (1/2)
