+++ lib/op-attrs/include/op-attrs/tensor_shape.h
@@ -16,7 +16,7 @@ struct TensorShape : public use_visitable_cmp<TensorShape> {
 
   template <typename Dims>
   TensorShape(Dims const &dims, DataType data_type)
-      : dims(this->dims), data_type(this->data_type) {}
+      : dims(dims), data_type(data_type) {}
 
   size_t at(ff_dim_t) const;
   size_t operator[](ff_dim_t) const;
