+++ lib/runtime/src/cuda_allocator.h
@@ -2,7 +2,6 @@
 #define _FLEXFLOW_RUNTIME_CUDA_ALLOCATOR_H
 
 #include "kernels/allocation.h"
-#include <cuda_runtime.h>
 
 namespace FlexFlow {
