+++ lib/runtime/src/task_spec/op_task_signature.h
@@ -41,7 +41,9 @@ struct OpTaskSignature {
   OpTaskSignature() = delete;
   explicit OpTaskSignature(OpTaskType);
 
-  OpTaskType get_task_type() const;
+  OpTaskType get_task_type() const {
+    return this->type;
+  }
 
   void add_input_slot(slot_id, SlotType slot_type = SlotType::TENSOR);
   void add_optional_input_slot(slot_id, SlotType slot_type = SlotType::TENSOR);
@@ -59,21 +61,17 @@ struct OpTaskSignature {
 
   void add_from_slot_spec(OpTensorSlotSpec const &spec);
 
-  /* void add_input_slot(slot_id, Legion::PrivilegeMode); */
-  /* void add_input_slot(slot_id, SlotType, Legion::PrivilegeMode); */
-
-  bool operator==(OpTaskSignature const &) const;
-  bool operator!=(OpTaskSignature const &) const;
-
   template <typename T>
   void add_arg_slot(slot_id name) {
     static_assert(is_serializable<T>::value, "Type must be serializable");
     this->task_arg_types.insert({name, type_index<T>()});
   }
 
-  // TODO: what is this function's purpose? OpTaskSignature doesn't need a return value
-  // template <typename T>
-  // void add_return_value();
+  // TODO: should this be a single type index? 
+  template <typename T>
+  void add_return_value() {
+    this->return_value.push_back(type_index<T>);
+  }
 
   // adds arg_slot without checking is_serializable, used for arguments that are
   // deviceSpecific
@@ -90,12 +88,13 @@ struct OpTaskSignature {
 
 private:
   OpTaskType type;
+  std::vector<std::type_index> return_value;
   std::unordered_map<slot_id, std::type_index> task_arg_types;
   std::unordered_set<OpTensorSlotSpec> op_tensor_slots;
 };
 
 OpTaskSignature get_op_signature(task_id_t const &);
-FF_VISITABLE_STRUCT_NONSTANDARD_CONSTRUCTION(OpTaskSignature, get_tensor_slots, set_arg_types, get_arg_types);
+FF_VISITABLE_STRUCT_NONSTANDARD_CONSTRUCTION(OpTaskSignature, op_tensor_slots, task_arg_types);
 
 template <typename F>
 void register_task(task_id_t,
